fix cl_app, refactor read_img
