+++ src/imgread/version.py
@@ -1,1 +1,1 @@
-__version__ = "0.1"
+__version__ = "0.1.1"
